# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/application-inventory/application": {
            "get": {
                "description": "List all applications.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "list"
                ],
                "summary": "List all applications.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Application"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create an application.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "create"
                ],
                "summary": "Create an application.",
                "parameters": [
                    {
                        "description": "Application data",
                        "name": "application",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Application"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Application"
                        }
                    }
                }
            }
        },
...
        "/application-inventory/application/{id}": {
            "get": {
                "description": "Get an application by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "Get an application by ID.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Application"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an application.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "Update an application.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Application id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Application data",
                        "name": "application",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Application"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "Delete an application.",
                "tags": [
                    "delete"
                ],
                "summary": "Delete an application.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Application id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
        "/application-inventory/application/{id}/bucket": {
            "get": {
                "description": "List all buckets.",
                "produces": [
                    "application/json"
                ],
                "summary": "List all buckets.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Bucket"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a bucket for an application.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a bucket for an application.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Bucket data",
                        "name": "bucket",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Bucket"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Bucket"
                        }
                    }
                }
            }
        },
...
        "/controls/bucket": {
            "get": {
                "description": "List all buckets.",
                "produces": [
                    "application/json"
                ],
                "summary": "List all buckets.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Bucket"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a bucket.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a bucket.",
                "parameters": [
                    {
                        "description": "Bucket data",
                        "name": "bucket",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Bucket"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Bucket"
                        }
                    }
                }
            }
        },
        "/controls/bucket/{id}": {
            "get": {
                "description": "Get a bucket by ID.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a bucket by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bucket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Bucket"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a bucket.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a bucket.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bucket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Bucket data",
                        "name": "bucket",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Bucket"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/api.Bucket"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a bucket.",
                "summary": "Delete a bucket.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bucket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/api.Bucket"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Application": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "businessService": {
                    "type": "string"
                },
                "comments": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "review": {
                    "$ref": "#/definitions/api.Review"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updateUser": {
                    "type": "string"
                }
            }
        },
        "api.Bucket": {
            "type": "object",
            "properties": {
                "application": {
                    "type": "integer"
                },
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "updateUser": {
                    "type": "string"
                }
            }
        },
...
        "api.Review": {
            "type": "object",
            "required": [
                "application"
            ],
            "properties": {
                "application": {
                    "type": "object",
                    "required": [
                        "id"
                    ],
                    "properties": {
                        "id": {
                            "type": "integer"
                        }
                    }
                },
                "businessCriticality": {
                    "type": "integer"
                },
                "comments": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "effortEstimate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "proposedAction": {
                    "type": "string"
                },
                "updateUser": {
                    "type": "string"
                },
                "workPriority": {
                    "type": "integer"
                }
            }
        },
...
        "model.Application": {
            "type": "object",
            "properties": {
                "businessService": {
                    "$ref": "#/definitions/model.BusinessService"
                },
                "businessServiceID": {
                    "type": "integer"
                },
                "comments": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "identities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Identity"
                    }
                },
                "name": {
                    "type": "string"
                },
                "review": {
                    "$ref": "#/definitions/model.Review"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Tag"
                    }
                },
                "updateUser": {
                    "type": "string"
                }
            }
        },
        "model.BusinessService": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/model.Stakeholder"
                },
                "ownerID": {
                    "type": "integer"
                },
                "updateUser": {
                    "type": "string"
                }
            }
        },
        "model.Identity": {
            "type": "object",
            "properties": {
                "applicationID": {
                    "type": "integer"
                },
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "encrypted": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "settings": {
                    "type": "string"
                },
                "updateUser": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
        "model.Review": {
            "type": "object",
            "properties": {
                "application": {
                    "$ref": "#/definitions/model.Application"
                },
                "applicationID": {
                    "type": "integer"
                },
                "businessCriticality": {
                    "type": "integer"
                },
                "comments": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "effortEstimate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "proposedAction": {
                    "type": "string"
                },
                "updateUser": {
                    "type": "string"
                },
                "workPriority": {
                    "type": "integer"
                }
            }
        },
        "model.Stakeholder": {
            "type": "object",
            "properties": {
                "businessServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BusinessService"
                    }
                },
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.StakeholderGroup"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "jobFunction": {
                    "$ref": "#/definitions/model.JobFunction"
                },
                "jobFunctionID": {
                    "type": "integer"
                },
                "updateUser": {
                    "type": "string"
                }
            }
        },
        "model.StakeholderGroup": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "stakeholders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Stakeholder"
                    }
                },
                "updateUser": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.Tag": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tagType": {
                    "$ref": "#/definitions/model.TagType"
                },
                "tagTypeID": {
                    "type": "integer"
                },
                "updateUser": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.TagType": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "rank": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Tag"
                    }
                },
                "updateUser": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}